Consider all peercertificates presented Move cert verification logic out of logger file
